+++ src/components/SidePanel/AnalogTriggerSettings.tsx
@@ -15,18 +15,11 @@ import {
 import { appState } from '../../slices/appSlice';
 import { getRecordingMode } from '../../slices/chartSlice';
 import {
-    getTriggerBias,
     getTriggerEdge,
-    getTriggerRecordingLength,
-    getTriggerType,
     getTriggerValue,
-    setTriggerBias,
     setTriggerEdge,
     setTriggerLevel,
-    setTriggerRecordingLength,
-    setTriggerType,
     TriggerEdgeValues,
-    TriggerTypeValues,
 } from '../../slices/triggerSlice';
 
 const CurrentUnitValues = ['mA', '\u00B5A'] as const;
@@ -50,15 +43,9 @@ const convertToMicroAmps = (unit: CurrentUnit, value: number) => {
     }
 };
 
-const calculateBiasTime = (recordingLength: number, bias: number) =>
-    Number((recordingLength * (bias / 100)).toFixed(2));
-
 export default () => {
     const dispatch = useDispatch();
-    const recordingLength = useSelector(getTriggerRecordingLength);
-    const triggerBias = useSelector(getTriggerBias);
     const triggerValue = useSelector(getTriggerValue);
-    const triggerType = useSelector(getTriggerType);
     const triggerEdge = useSelector(getTriggerEdge);
     const { samplingRunning } = useSelector(appState);
     const dataLoggerActive =
@@ -73,12 +60,6 @@ export default () => {
 
     const [internalTriggerValue, setInternalTriggerValue] =
         useState(triggerValue);
-    const [internalTriggerLength, setInternalTriggerLength] =
-        useState(recordingLength);
-    const [triggerBiasValue, setTriggerBiasValue] = useState(triggerBias);
-    const [computedBias, setComputedBias] = useState(
-        calculateBiasTime(internalTriggerLength, triggerBias)
-    );
 
     useEffect(() => {
         if (triggerValue > 1000) {
@@ -90,62 +71,8 @@ export default () => {
         }
     }, [triggerValue]);
 
-    useEffect(() => {
-        setInternalTriggerLength(recordingLength);
-    }, [recordingLength]);
-
-    useEffect(() => {
-        setTriggerBiasValue(triggerBias);
-    }, [triggerBias]);
-
     return (
         <>
-            <NumberInput
-                range={{
-                    min: 1,
-                    max: 1000,
-                    decimals: 2,
-                    step: 0.01,
-                }}
-                title="Duration of trigger window"
-                value={internalTriggerLength}
-                onChange={setInternalTriggerLength}
-                onChangeComplete={(value: number) => {
-                    dispatch(setTriggerRecordingLength(value));
-                    setComputedBias(calculateBiasTime(value, triggerBias));
-                }}
-                unit="ms"
-                label="Length"
-                disabled={dataLoggerActive}
-                showSlider
-            />
-            <div className="tw-flex tw-flex-col tw-gap-2">
-                <NumberInput
-                    range={{
-                        min: 0,
-                        max: 100,
-                        decimals: 0,
-                        step: 1,
-                    }}
-                    title='Trigger bias from "Start of trigger window"'
-                    value={triggerBiasValue}
-                    onChange={setTriggerBiasValue}
-                    onChangeComplete={(value: number) => {
-                        dispatch(setTriggerBias(value));
-                        setComputedBias(
-                            calculateBiasTime(internalTriggerLength, value)
-                        );
-                    }}
-                    unit="%"
-                    label="Bias"
-                    disabled={dataLoggerActive}
-                    showSlider
-                />
-                <div className="tw-text-sm tw-text-gray-500">
-                    Computed bias: {computedBias} ms
-                </div>
-            </div>
-
             <NumberInput
                 range={{
                     min: getMin(levelUnit),
@@ -178,12 +105,6 @@ export default () => {
                 label="Level"
                 disabled={dataLoggerActive}
                 showSlider
-            />
-            <StateSelector
-                items={[...TriggerTypeValues]}
-                onSelect={m => dispatch(setTriggerType(TriggerTypeValues[m]))}
-                selectedItem={triggerType}
-                disabled={samplingRunning}
             />
             <StateSelector
                 items={[...TriggerEdgeValues]}